Integrate new version of the DirectX Shader Compiler

+++ D3D12/Graphics/D3DUtils.h
@@ -42,6 +42,6 @@ inline void SetD3DObjectName(ID3D12Object* pObject, const char* pName)
 {
 	if (pObject)
 	{
-		HR(pObject->SetPrivateData(WKPDID_D3DDebugObjectName, strlen(pName), pName));
+		HR(pObject->SetPrivateData(WKPDID_D3DDebugObjectName, (uint32)strlen(pName), pName));
 	}
 }
